Extract Sync logic into SyncRepository
